refactor: implement clearable input functionality across control components
