+++ packages/combinator/src/components/controls/control_string.vue
@@ -49,7 +49,8 @@
 import { DtInput, DtText, VALIDATION_MESSAGE_TYPES } from '@dialpad/dialtone-vue';
 import DtcControlClearableShell from './control_clearable_shell.vue';
 import { VALUE_UPDATE_EVENT } from '@/src/lib/constants';
-import { computed, nextTick, ref, watch } from 'vue';
+import { computed } from 'vue';
+import { useClearableInput } from '@/src/lib/utils_vue';
 
 const props = defineProps({
   value: {
@@ -80,16 +81,8 @@ const props = defineProps({
 
 const emit = defineEmits([VALUE_UPDATE_EVENT]);
 
-const expanded = ref(false);
-const inputRef = ref(null);
-const hasPendingValue = ref(false);
-const hasInternalUpdate = ref(false);
-
-const inputValue = computed(() => props.value ?? '');
-
-const isEmpty = computed(() => props.value === null || props.value === undefined || props.value === '');
-
-const clearDisabled = computed(() => !props.clearable || props.required || props.disabled || isEmpty.value);
+const { expanded, inputRef, inputValue, isEmpty, updateValue, addValue, collapseIfEmpty, clearValue } =
+  useClearableInput({ props, emit });
 
 const messages = computed(() => {
   const messages = [];
@@ -100,40 +93,6 @@ const messages = computed(() => {
     });
   }
   return messages;
-});
-
-function updateValue (value) {
-  expanded.value = true;
-  hasInternalUpdate.value = true;
-  hasPendingValue.value = value !== null && value !== undefined && value !== '';
-  emit(VALUE_UPDATE_EVENT, value);
-}
-
-async function addValue () {
-  expanded.value = true;
-  await nextTick();
-  inputRef.value?.focus();
-}
-
-function collapseIfEmpty () {
-  if (!isEmpty.value || hasPendingValue.value) return;
-  expanded.value = false;
-}
-
-function clearValue () {
-  if (clearDisabled.value) return;
-  expanded.value = false;
-  hasPendingValue.value = false;
-  emit(VALUE_UPDATE_EVENT, null);
-}
-
-watch(() => props.value, () => {
-  hasPendingValue.value = false;
-  if (hasInternalUpdate.value) {
-    hasInternalUpdate.value = false;
-    return;
-  }
-  if (isEmpty.value) expanded.value = false;
 });
 </script>
